+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 CHANGELOG
 ---------
+- **2017-09-27**
+  - Add `timeutil.GetFormat`
 - **2017-09-24**
   - Add `phonenumber.AreaCodeToGeo`
   - Add `timeutil.TimeRange`